+++ src/mcp/tool.ts
@@ -7,6 +7,7 @@ import { RC } from "../rc";
 
 export interface ServerToolContext {
   projectId?: string;
+  accountEmail?: string | null;
   config: Config;
   host: FirebaseMcpServer;
   rc: RC;
@@ -29,6 +30,8 @@ export interface ServerTool<InputSchema extends ZodTypeAny = ZodTypeAny> {
       requiresProject?: boolean;
       /** Set this on a tool if it *always* requires a signed-in user to work. */
       requiresAuth?: boolean;
+      /** Tools are grouped by feature. --only can configure what tools is available. */
+      feature?: string;
     };
   };
   fn: (input: z.infer<InputSchema>, ctx: ServerToolContext) => Promise<CallToolResult>;